+++ astropy/coordinates/builtin_frames/utils.py
@@ -125,6 +125,69 @@ def norm(p):
     return p/np.sqrt(np.einsum('...i,...i', p, p))[..., np.newaxis]
 
 
+def prepare_earth_position_vel(time):
+    """
+    Get barycentric position and velocity, and heliocentric position of Earth
+
+    Parameters
+    -----------
+    time : `~astropy.time.Time`
+        time at which to calculate position and velocity of Earth
+
+    Returns
+    --------
+    earth_pv : `np.ndarray`
+        Barycentric position and velocity of Earth, in au and au/day
+    earth_helio : `np.ndarray`
+        Heliocentric position of Earth in au
+    """
+    # this goes here to avoid circular import errors
+    from ..solar_system import get_body_barycentric, solar_system_ephemeris
+    # get barycentric position and velocity of earth
+    ephemeris = solar_system_ephemeris.get()
+    earth_pv = get_body_barycentric('earth', time,
+                                    ephemeris=ephemeris, get_velocity=True)
+
+    # get heliocentric position of earth
+    sun = get_body_barycentric('sun', time, ephemeris=ephemeris)
+    earth_heliocentric = (earth_pv[0].xyz - sun.xyz).to(u.au)
+
+    # prepare to pass to erfa
+    earth_pv = np.array([earth_pv[0].xyz.to(u.au), earth_pv[1].xyz.to(u.au/u.d)])
+    earth_pv = np.rollaxis(np.rollaxis(earth_pv, 0, earth_pv.ndim), 0, earth_pv.ndim)
+    earth_heliocentric = np.rollaxis(earth_heliocentric, 0, earth_heliocentric.ndim)
+    return earth_pv, earth_heliocentric
+
+
+def get_cip(jd1, jd2):
+    """
+    Find the X, Y coordinates of the CIP and the CIO locator, s.
+
+    Parameters
+    ----------
+    jd1 : float or `np.ndarray`
+        First part of two part Julian date (TDB)
+    jd2 : float or `np.ndarray`
+        Second part of two part Julian date (TDB)
+
+    Returns
+    --------
+    x : float or `np.ndarray`
+        x coordinate of the CIP
+    y : float or `np.ndarray`
+        y coordinate of the CIP
+    s : float or `np.ndarray`
+        CIO locator, s
+    """
+    # classical NPB matrix, IAU 2006/2000A
+    rpnb = erfa.pnm06a(jd1, jd2)
+    # CIP X, Y coordinates from array
+    x, y = erfa.bpn2xy(rpnb)
+    # CIO locator, s
+    s = erfa.s06(jd1, jd2, x, y)
+    return x, y, s
+
+
 def aticq(ri, di, astrom):
     """
     A slightly modified version of the ERFA function ``eraAticq``.
